Move `Combine` helpers into a new file
